extract KubectlError, KubectlLogLine, KubectlLogResult into k8s_types.py
Move shared data types out of kubectl.py into a dedicated k8s_types module
to break the import dependency on the concrete Kubectl implementation.
All imports updated across production and test code.
