+++ lib/iris/src/iris/cluster/k8s/k8s_service_impl.py
@@ -12,7 +12,7 @@ from __future__ import annotations
 import subprocess
 from datetime import UTC, datetime
 
-from iris.cluster.k8s.kubectl import KubectlError, KubectlLogLine, KubectlLogResult
+from iris.cluster.k8s.k8s_types import KubectlError, KubectlLogLine, KubectlLogResult
 from iris.cluster.service_mode import ServiceMode
 
 # Resource types that K8s recognizes in container resource requests/limits.
