+++ lib/iris/tests/cluster/k8s/test_k8s_service_impl.py
@@ -9,7 +9,7 @@ import pytest
 
 from iris.cluster.k8s.k8s_service import K8sService
 from iris.cluster.k8s.k8s_service_impl import K8sServiceImpl
-from iris.cluster.k8s.kubectl import KubectlError
+from iris.cluster.k8s.k8s_types import KubectlError
 
 
 def _pod_manifest(
